requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.5
dnspython==2.3.0
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
Werkzeug==2.3.6
